Move Horde and SD endpoints into separate files
